Refactored logic into deckService as it made more sense
