+++ src/main/java/com/blackjack/repository/CardRepository.java
@@ -9,5 +9,5 @@ import com.blackjack.models.Card;
  * It extends the JpaRepository interface, providing additional methods for working with cards.
  */
 public interface CardRepository extends JpaRepository<Card, Long>{
-	boolean existsAny();
+	boolean existsBy();
 }
